+++ CHANGELOG.md
@@ -24,10 +24,12 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ### Changed
 
 - `New container` task interface to default for edit and new
+- Only primary type type materials are displayed as 'Type material' in distribution maps [#4567]
 
 [#3132]: https://github.com/SpeciesFileGroup/taxonworks/issues/3132
 [#3154]: https://github.com/SpeciesFileGroup/taxonworks/issues/3154
 [#4557]: https://github.com/SpeciesFileGroup/taxonworks/issues/4557
+[#4567]: https://github.com/SpeciesFileGroup/taxonworks/issues/4567
 
 ## [0.54.0] - 2025-09-24